backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
pymongo>=4.7,<5.0
PyJWT>=2.8,<3.0
werkzeug>=3.0,<4.0
python-dotenv>=1.0,<2.0
cloudinary>=1.40,<2.0
pdfplumber>=0.11,<1.0
python-docx>=1.1,<2.0
google-genai>=1.0,<2.0
